Rewrite 'tileset-to-array.py' to greatly simplify its use
The script interface is entirely new. The input image doesn't need to
follow a strict format anymore and it should only contain the tileset
data. The palette is now specified in a separate file and the tile size
is passed to the script through options.

Plus, the script can now generate image arrays of 8bpp format.