Simplify scene initialization code
Previously, 'scene_set' took a 'flags' argument which determined if the
'init' function of a scene should be called. The idea was that all
information should be passed to the 'init' function as bits in the
'flags' argument.

Now, it is possible to pass any data structure to 'scene_set'. The
'SCENE_SET' macro is a shorthand for setting a scene without calling its
'init' function.