Add 'ARM' function attribute macro
By default the template generates code for 'ARM'. For this reason, the
'THUMB' attribute was already present but the 'ARM' attribute was
missing.

However, if the developer decides to set 'THUMB' as the default (e.g. by
changing the 'CFLAGS' variable in the Makefile), then the 'ARM'
attribute will be needed.