Quality: rename 'Entity' structure into 'entity_Type'
That structure contains constants and functions specific to a type of
entity (like Player, Enemy...) while 'entity_Data' stores the state of
each entity. This is similar to classes and instances in OOP.

The name 'Entity' was therefore too generic, given that the structure
represented only the type and not also the actual object.

The 'entity_type' function was also renamed to 'entity_get_type' to
better avoid naming collision with 'entity_Type' variables.