Scene: add 'scene_transition_to' function and transition scene
This special scene creates transitions between the current scene and any
next scene. As is, it uses the display's 'darken' graphical effect, but
it can easily be made to use other effects (such as mosaic, lighten...).

The transition time is set to 64 ticks (slightly more than a second),
but that can also be immediately changed.